requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
ezdxf
numpy
Pillow
python-dotenv
